refactor(global): 전역으로 사용되는 파일 이동
