+++ src/main/java/org/sopt/global/error/BusinessException.java
@@ -1,4 +1,4 @@
-package org.sopt.common.exception;
+package org.sopt.global.error;
 
 public class BusinessException extends RuntimeException {
