+++ src/main/java/org/sopt/global/error/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package org.sopt.common.exception;
+package org.sopt.global.error;
 
-import org.sopt.common.response.ErrorResponse;
+import org.sopt.global.common.response.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
