+++ src/main/java/org/sopt/global/util/GraphemeClusterUtil.java
@@ -1,4 +1,4 @@
-package org.sopt.common.util;
+package org.sopt.global.util;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
